+++ examples/internal-routing/batfish.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 
 module.exports = () => {
   return {
-    wrapperPath: path.join(__dirname, './src/wrapper.js'),
+    applicationWrapperPath: path.join(__dirname, './src/wrapper.js'),
     siteOrigin: 'https://www.mapbox.com'
   };
 };