+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Head
 
+-   [Breaking change] Changed `wrapperPath` option to `applicationWrapperPath`.
 -   [Breaking change] Removed `notFoundPath` option.
     Instead, 404 pages always live as `404.(js|md)` in the `pagesDirectory`.
 -   [Breaking change] 404 pages are no longer `index.html` files inside directories, e.g. `404/index.html`.